+++ agents-flex-store/agents-flex-store-aliyun/src/main/java/com/agentsflex/store/aliyun/AliyunVectorStore.java
@@ -203,12 +203,11 @@ public class AliyunVectorStore extends DocumentStore {
             document.setVector(jsonObject.getObject("vector", float[].class));
             // 阿里云数据采用余弦相似度计算 jsonObject.getDoubleValue("score") 表示余弦距离，
             // 原始余弦距离范围是[0, 2]，0表示最相似，2表示最不相似
-            Double distance = jsonObject.getDouble("score");
+            Float distance = jsonObject.getFloat("score");
             if (distance != null) {
-                double score = distance / 2.0;
-                document.setScore(1.0d - score);
+                float score = distance / 2.0f;
+                document.setScore(1.0f - score);
             }
-
 
             JSONObject fields = jsonObject.getJSONObject("fields");
             document.addMetadata(fields);
